fix lumnace setting for grating, sinusoidal mode. Stim contrast in Gabor mode is not correct, need to be fixed

+++ PTB/Stim/Make_BarOrientationTexture.m
@@ -5,7 +5,7 @@ bar_h = sobj.RECT(4);
 bar_w = sobj.StimSize_pix(1);
 
 %Texture objext
-im_matrix = ones(bar_h, bar_w) * sobj.stimlumi;
+im_matrix = uint8(ones(bar_h, bar_w)) * sobj.stimlumi;
 im_tex = Screen('MakeTexture', sobj.wPtr, im_matrix, sobj.BarOrientation);
 
 end
